Gracefully handle registered saved games in shareware version
Once we add the ability to switch game paths, it's possible to get into a
situation where the user profile has registered version saved games, but
the current game path is pointing to a shareware version. Trying to load
one of these saved games would then throw an exception (due to the
level file not being present in the data) and quit the game.

To avoid this, check if the slot can be loaded before triggering the load.
For simplicity, we "misuse" the message that's shown when trying to start
a new game in episode 2 to 4 from a shareware version. This saves us
from having to create a message script programmatically. In the future,
it would be nicer to have a more fitting message however.